changed the class directories and created a separate package for them currency
