+++ src/main/java/org/example/currency/eur/RateEUR.java
@@ -1,4 +1,4 @@
-package org.example.eur;
+package org.example.currency.eur;
 
 public interface RateEUR {
     double sellEUR(double amountEur);
