+++ src/main/java/org/example/currency/usd/RateUSD.java
@@ -1,4 +1,4 @@
-package org.example.usd;
+package org.example.currency.usd;
 
 public interface RateUSD {
     double sellUSD(double amountUsd);
